Refactor : TodoOtherList.jsx 파일 분리
