+++ src/Pages/TodoPage/OtherCadetList.jsx
@@ -2,24 +2,13 @@ import React, { useState, useEffect} from 'react';
 import { TodoService } from 'Network';
 import { format } from 'date-fns';
 
+import TodoOtherList from './TodoOtherList';
 import { checkDateTodo } from 'Utils';
 
 import styled from 'styled-components';
 
-import Checkbox from '@mui/material/Checkbox';
-
 const OtherTitle = (() => {
   return (<OtherTitleName>👀 다른 카뎃은 무엇을?</OtherTitleName>);
-})
-
-const WarningSignNoList = (() => {
-  return (
-    <WarningSignList>
-      <span style={{ color: 'gray' }}>
-       등록된 할 일이 없습니다!
-      </span>
-    </WarningSignList>
-  );
 })
 
 const WarningSignNotVaildDate = (() => {
@@ -28,48 +17,6 @@ const WarningSignNotVaildDate = (() => {
       아직 진행하지 않은 날짜입니다!
     </WarningSignDate>
   )
-})
-
-const OtherName = (({index, item}) => {
-  return (
-    <OtherNames key={index}>
-      {item.userName}
-    </OtherNames>
-  );
-})
-
-const OtherListCheckbox = (({list}) => {
-  return (
-    <OtherListCheckboxes>
-      <Checkbox
-        checked={list.titleCheck}
-        size="small"
-        sx={{
-          '& .MuiSvgIcon-root': { fontSize: 15 },
-        }}
-        disabled
-      />
-  </OtherListCheckboxes>
-  );
-})
-
-const OtherListObject = (({list}) => {
-  return (
-    <span>
-      {list.titleCheck === true ? (
-          <span
-            style={{
-              textDecorationLine: 'line-through',
-              color: 'gray',
-          }}
-        >
-          {list.title}
-        </span>
-      ) : (
-          <span>{list.title}</span>
-      )}
-    </span>
-  );
 })
 
 const OtherCadetList = ({date}) => {
@@ -89,29 +36,9 @@ const OtherCadetList = ({date}) => {
     <TodoOtherBody>
       <OtherTitle />
       {checkDateTodo(date) ? (
-      <WarningSignNotVaildDate />
+        <WarningSignNotVaildDate />
       ) : (
-        <TodoOtherListBody key="index">
-        {othersToDo.map((item, index) => (
-          <TodoOtherListContainer key={index}>
-            <OtherName index={index} item={item}/>
-            <TodoEachListBody>
-              {item.todoDtoList.length === 0 ? (
-                <WarningSignNoList />
-              ) : (
-                <TodoEachListContainer>
-                  {item.todoDtoList.map((list, index) => (
-                    <TodoEachListEntity key={index}>
-                      <OtherListCheckbox list={list}/>
-                      <OtherListObject list={list}/>
-                    </TodoEachListEntity>
-                  ))}
-                </TodoEachListContainer>
-              )}
-            </TodoEachListBody>
-          </TodoOtherListContainer>
-        ))}
-          </TodoOtherListBody>
+        <TodoOtherList othersToDo={othersToDo} />
       )}
     </TodoOtherBody>
   );
@@ -128,33 +55,6 @@ flex-direction: column;
 width: 50%;
 height: 100%;
 `
-const TodoOtherListBody = styled.div`
-display: grid;
-grid-template-columns: 1fr 1fr;
-gap: 10px;
-padding: 10px;
-margin-bottom: 10px;
-`
-const TodoOtherListContainer = styled.div`
-border: 0px;
-padding: 1em;
-border-radius: 1em;
-background-color: white;
-width: 100%;
-box-sizing: border-box;
-height: 300px;
-`
-const TodoEachListBody = styled.div`
-overflow: auto;
-margin-top: 5%;
-margin-left: 5%;
-height: 80%;
-`
-const TodoEachListContainer = styled.div`
-`
-const TodoEachListEntity = styled.div`
-display: table;
-`
 const OtherTitleName = styled.div`
 font-size: 25px;
 margin-bottom: 10px;
@@ -165,20 +65,6 @@ align-item: center;
 text-align: center;
 margin-top: 43%;
 margin-bottom: 40%;
-`
-const WarningSignList = styled.div`
-text-align: center;
-// vertical-align: middle;
-margin-top: 80px;
-margin-left: -5%;
-`
-const OtherNames = styled.div`
-margin-top: -5px;
-font-size: 20px;
-`
-const OtherListCheckboxes = styled.span`
-margin-left: -10px;
-margin-bottom: -5px;
 `
 
 export default React.memo(OtherCadetList);
